# [SPARK-16360][SQL] Speed up SQL query performance by removing redundant `executePlan` call