@hvanhovell 2016-07-04
Any idea what causes the regression? 5 seconds seems way too long for analysis...