@dongjoon-hyun 2016-07-04
Thank you for review, @hvanhovell . BTW, it's over 12 seconds for one single analysis.

Elapsed time: 25.787751452s  --> Elapsed time: 12.364812255s.

The reason I executed `time(sql(query))` two times is that SQL parser and other overhead.