@dongjoon-hyun 2016-07-04
Oh, thank you for advice of `dumpTimeSpent`. I didn't look at in that way.
In these days, I'm trying to investigate large queries situation.
This analysis is very helpful for me. Thank you so much, @hvanhovell .