@dongjoon-hyun 2016-07-04
Interesting result. We definitely need to take a look at `ResolveReferences`-related stuff.