@dongjoon-hyun 2016-07-04
Yep. I agree.
Could you make a PR for that? I think we also have some optimization points about that.