@dongjoon-hyun 2016-07-05
Thank you for review, @liancheng .
I'm sure that the performance of Analyzer need to be improved. But, in any cases, the cost of analyzer cannot be zero.
We should skip the redundant analysis. IMO, that idea sounds orthogonal to this PR. So, I asked @hvanhovell to make a PR for that.