@dongjoon-hyun 2016-07-05
Hi, @cloud-fan , @hvanhovell , @liancheng .

According to @cloud-fan 's advice, after changing the following, it turns out that the difference is not noticeable.
```
-    new Dataset[Row](sparkSession, qe, RowEncoder(qe.analyzed.schema), skipAnalysis = true)
+    new Dataset[Row](sparkSession, qe, RowEncoder(qe.analyzed.schema))
```

Exactly as you guys told, the second call of `qe.assertAnalyzed()` is not the root cause. The only difference resides on `sparkSession.sessionState.executePlan(logicalPlan)`.

I'll update the PR.