@dongjoon-hyun 2016-07-05
Now, I update the title and description of PR/JIRA.
The only patch in this PR is the following one word change.
```
-    new Dataset[Row](sparkSession, logicalPlan, RowEncoder(qe.analyzed.schema))
+    new Dataset[Row](sparkSession, qe, RowEncoder(qe.analyzed.schema))
```
Thank you all for fast review & advice. At first commit, I thought it is important to remove all repeating logics. But, now only the minimum meaningful code change remains.