@liancheng 2016-07-05
Merged to master. Thanks!