@dongjoon-hyun 2016-07-05
Thank you for review and merging, @liancheng , @cloud-fan , @hvanhovell , and @naliazheli !